\name{dvonmises}
\alias{dvonmises}
\title{Calculated the VonMises Density.}
\usage{
dvonmises(y, mu, kappa, log = FALSE)
}
\arguments{
\item{y}{Observations}

\item{mu}{Parameter mu}

\item{kappa}{Parameter kappa}

\item{log}{Logical; if TRUE, probabilities p are given as log(p)}
}
\value{
VonMises Density
}
\description{
Calculated the VonMises Density.
}
