\name{print.circmax}
\alias{print.circmax}
\title{Print output of a circmax object.}
\usage{
\method{print}{circmax}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{Circmax object.}

\item{digits}{Number of digits.}

\item{...}{Additional parameters.}
}
\description{
Print output of a circmax object.
}
