\name{vonmises_shinyapp}
\alias{vonmises_shinyapp}
\title{Shiny app for the Von Mises Distribution.}
\usage{
  vonmises_shinyapp(\dots)
}
\arguments{
  \item{\dots}{Additional parameters passed on to \code{\link[shiny]{runApp}}.}
}
\description{
Runs a shiny app for visualization of the Von Mises Distribution.
}
