\name{coef.circmax}
\alias{coef.circmax}
\title{Print coefficients of a circmax object.}
\usage{
\method{coef}{circmax}(object, model = c("full", "location", "scale"), ...)
}
\arguments{
\item{object}{Circmax object.}

\item{model}{Model (full, location or scale).}

\item{...}{Additional parameters.}
}
\description{
Print coefficients of a circmax object.
}
